+++ R/validate_inputs.R
@@ -107,9 +107,12 @@ do_validate_population <- function(population) {
 #' @param shape A file object (path, hash, filename) corresponding to
 #'   the input shape file.
 #'
+#' @param strict If FALSE then run less stringent validation rules, used
+#'   for data exploration mode.
+#'
 #' @return An error if invalid.
 #' @keywords internal
-do_validate_programme <- function(programme, shape) {
+do_validate_programme <- function(programme, shape, strict = TRUE) {
   assert_file_extension(programme, "csv")
   data <- read_csv(programme$path, header = TRUE)
   data$art_current <- as.numeric(data$art_current)
@@ -118,10 +121,9 @@ do_validate_programme <- function(programme, shape) {
     colnames(data),
     c("area_id", "calendar_quarter", "sex", "age_group", "art_current"))
   shape_regions <- read_regions(shape, "shape")
-  programme_regions <- read_regions(programme, "programme")
-  assert_consistent_regions(shape_regions$area_id, programme_regions,
+  assert_consistent_regions(shape_regions$area_id, unique(data$area_id),
                             "programme")
-  assert_single_level(shape_regions, programme_regions)
+  assert_single_level_per_year(shape_regions, data)
   assert_unique_combinations(
     data, c("area_id", "calendar_quarter", "sex", "age_group"))
   assert_expected_values(data, "sex", c("male", "female", "both"))
@@ -146,9 +148,12 @@ do_validate_programme <- function(programme, shape) {
 #' @param shape A file object (path, hash, filename) corresponding to
 #'   the input shape file.
 #'
+#' @param strict If FALSE then run less stringent validation rules, used
+#'   for data exploration mode.
+#'
 #' @return An error if invalid.
 #' @keywords internal
-do_validate_anc <- function(anc, shape) {
+do_validate_anc <- function(anc, shape, strict = TRUE) {
   assert_file_extension(anc, "csv")
   data <- read_csv(anc$path, header = TRUE)
   assert_single_country(data, "anc")
@@ -157,16 +162,16 @@ do_validate_anc <- function(anc, shape) {
     c("area_id", "age_group", "year", "anc_clients",
       "anc_known_pos", "anc_already_art", "anc_tested", "anc_tested_pos"))
   shape_regions <- read_regions(shape, "shape")
-  anc_regions <- read_regions(anc, "anc")
-  assert_consistent_regions(shape_regions$area_id, anc_regions,
-                            "ANC")
-  assert_single_level(shape_regions, anc_regions)
+  assert_consistent_regions(shape_regions$area_id, unique(data$area_id), "ANC")
+  assert_single_level_per_year(shape_regions, data)
   assert_unique_combinations(data, c("area_id", "age_group", "year"))
   assert_expected_values(data, "age_group", "Y015_049")
   assert_year_column(data)
   assert_column_positive_numeric(data, c("anc_clients", "anc_known_pos", "anc_already_art",
                                          "anc_tested", "anc_tested_pos"))
-  assert_anc_client_numbers(data)
+  if (strict) {
+    assert_anc_client_numbers(data)
+  }
   data <- naomi::calculate_prevalence_art_coverage(data)
   list(data = data,
        filters = list("year" = get_year_filters(data),
@@ -179,9 +184,14 @@ do_validate_anc <- function(anc, shape) {
 #'
 #' @param survey Path to input survey file.
 #'
+#' @param shape Path to input shape file.
+#'
+#' @param strict If FALSE then run less stringent validation rules, used
+#'   for data exploration mode.
+#'
 #' @return An error if invalid.
 #' @keywords internal
-do_validate_survey <- function(survey, shape) {
+do_validate_survey <- function(survey, shape, strict = TRUE) {
   assert_file_extension(survey, "csv")
   data <- read_csv(survey$path, header = TRUE)
   assert_single_country(data, "survey")